Moving the logic for getting a single policy in a different controller
